// src/modules/vehicle-intelligence-graph/domain/signal/graph-signal-type.ts
export enum GraphSignalType {
  MAINTENANCE_RISK_SIGNAL = 'MAINTENANCE_RISK_SIGNAL',
  COMPONENT_HEALTH_SIGNAL = 'COMPONENT_HEALTH_SIGNAL',
  SERVICE_DEPENDENCY_SIGNAL = 'SERVICE_DEPENDENCY_SIGNAL',
  ANOMALY_DETECTION_SIGNAL = 'ANOMALY_DETECTION_SIGNAL',
  TRUST_CONFLICT_SIGNAL = 'TRUST_CONFLICT_SIGNAL',
  INTELLIGENCE_ALERT_SIGNAL = 'INTELLIGENCE_ALERT_SIGNAL',
}